userland: remove spawn syscall
Spawn struggles to be a fork+exec replacement. In POSIX, we are
used (required?) to doing things between fork and exec for
spawning processes in all but the most trivial cases, and the
current spawn API is way too simplistic for use. Not only that,
but the posix_spawn API cannot be used to replace it due to
incredible complexity and poor extendability.

In the future, Ironclad will move to a Linux-like clone with
support for CLONE_VFORK to increase speed. For now, removing
spawn is the step to do.

Signed-off-by: streaksu <streaksu@mailbox.org>