devices/pci: add e1000 driver (#24)
Adds a (very basic, tested only on QEMU) E1000 driver. Has the same issues as the RTL8139 driver (doesn't read out the packets in the ISR, for example, which is more of a problem on real hardware on this card since it has a gigabit model). More handling will be needed to properly init a real-hardware card, especially given the amount of different models which need slightly different register values in places.

Other than that, its IMO in a good-enough shape to merge into upstream.

Reviewed-on: https://codeberg.org/Ironclad/Ironclad/pulls/24
Co-authored-by: Alexander Richards <electrodeyt@gmail.com>
Co-committed-by: Alexander Richards <electrodeyt@gmail.com>